Revert "Amend module"
