+++ backend/lesson/serializers.py
@@ -85,17 +85,21 @@ class LessonSerializer(ModelSerializer):
             raise ValidationError(f"Github URL musi zaczynać się od {GITHUB_REPO}.")
         return github_url
 
-    def create(self, validated_data):
-        technologies = validated_data.pop("technologies", [])
-        lesson = Lesson.objects.create(**validated_data)
+    def add_technologies(self, lesson, technologies):
         lesson.technologies.set(technologies)
+        return lesson
+
+    def create(self, validated_data):
+        technologies = validated_data.pop("technologies")
+        lesson = Lesson.objects.create(**validated_data)
+        self.add_technologies(lesson, technologies)
 
         if lesson.active:
             notify_lecturer(lesson)
 
         return lesson
 
-    def update(self, instance: Lesson, validated_data):
+    def update(self, instance, validated_data):
         technologies = validated_data.pop("technologies", [])
         current_price = instance.price
         new_price = validated_data.get("price", current_price)
@@ -106,8 +110,8 @@ class LessonSerializer(ModelSerializer):
         for attr, value in validated_data.items():
             setattr(instance, attr, value)
 
-        instance.technologies.set(technologies)
         instance.save()
+        instance.technologies.set(technologies)
 
         if instance.active:
             notify_lecturer(instance)
