+++ backend/module/serializers.py
@@ -3,6 +3,16 @@ from module.models import Module
 from lesson.models import Lesson
 
 
+def get_module_lessons(module):
+    module_lessons = (
+        Module.lessons.through.objects.filter(module=module).all().order_by("id")
+    )
+    return [
+        Lesson.objects.get(id=module_lesson.lesson_id)
+        for module_lesson in module_lessons
+    ]
+
+
 class LessonSerializer(ModelSerializer):
     class Meta:
         model = Lesson
@@ -22,8 +32,8 @@ class ModuleGetSerializer(ModelSerializer):
             "modified_at",
         )
 
-    def get_lessons(self, module: Module):
-        return LessonSerializer(module.ordered_lessons, many=True).data
+    def get_lessons(self, module):
+        return LessonSerializer(get_module_lessons(module=module), many=True).data
 
 
 class ModuleSerializer(ModelSerializer):
@@ -34,20 +44,29 @@ class ModuleSerializer(ModelSerializer):
             "modified_at",
         )
 
+    def add_lesson(self, module, lessons):
+        for lesson in lessons:
+            module.lessons.add(lesson)
+
+        return module
+
     def create(self, validated_data):
-        lessons = validated_data.pop("lessons", [])
+        lessons = validated_data.pop("lessons")
 
         module = Module.objects.create(**validated_data)
-        module.lessons.set(lessons)
+        module = self.add_lesson(module=module, lessons=lessons)
         module.save()
 
         return module
 
-    def update(self, instance: Module, validated_data):
-        lessons = validated_data.pop("lessons", [])
+    def update(self, instance, validated_data):
+        lessons = validated_data.pop("lessons")
+
         Module.objects.filter(pk=instance.pk).update(**validated_data)
 
+        instance = Module.objects.get(pk=instance.pk)
         instance.lessons.clear()
-        instance.lessons.set(lessons)
+        instance = self.add_lesson(module=instance, lessons=lessons)
+        instance.save()
 
         return instance
