+++ backend/profile/lecturers/views.py
@@ -16,10 +16,8 @@ from config_global import DUMMY_LECTURER_EMAIL
 
 class LecturerViewSet(ModelViewSet):
     http_method_names = ["get"]
-    queryset = (
-        LecturerProfile.objects.all()
-        .exclude(Q(title__isnull=True) | Q(description__isnull=True))
-        .order_by("id")
+    queryset = LecturerProfile.objects.all().exclude(
+        Q(title__isnull=True) | Q(description__isnull=True)
     )
     serializer_class = LecturerSerializer
     filterset_class = LecturerFilter
@@ -58,13 +56,9 @@ class LecturerViewSet(ModelViewSet):
 
 class BestLecturerViewSet(ModelViewSet):
     http_method_names = ["get"]
-    queryset = (
-        LecturerProfile.objects.exclude(
-            Q(title__isnull=True) | Q(description__isnull=True)
-        )
-        .all()
-        .order_by("id")
-    )
+    queryset = LecturerProfile.objects.exclude(
+        Q(title__isnull=True) | Q(description__isnull=True)
+    ).all()
     serializer_class = BestLecturerSerializer
 
     def get_rating(self, queryset):
