+++ backend/profile/logout/views.py
@@ -8,7 +8,7 @@ from django.contrib.auth import logout
 
 class ProfileLogoutViewSet(ModelViewSet):
     http_method_names = ["post"]
-    queryset = User.objects.all().order_by("id")
+    queryset = User.objects.all()
     serializer_class = ProfileLogoutSerializer
 
     def create(self, request, *args, **kwargs):
