+++ backend/tests/test_module.py
@@ -287,374 +287,3 @@ class ModuleTest(APITestCase):
         lessons_ids = data.pop("lessons")
         self.assertTrue(is_data_match(get_module(data["id"]), data))
         self.assertTrue(lessons_ids, get_module(data["id"]).lessons)
-
-
-class ModuleFilterTest(APITestCase):
-    def setUp(self):
-        self.endpoint = "/api/modules"
-        self.admin_data = {
-            "email": "admin_test_email@example.com",
-            "password": "TestPassword123",
-        }
-        self.admin_user = create_user(
-            first_name="first_name",
-            last_name="last_name",
-            email=self.admin_data["email"],
-            password=self.admin_data["password"],
-            is_active=True,
-            is_staff=True,
-        )
-        self.admin_profile = create_admin_profile(
-            profile=create_profile(user=self.admin_user, user_type="A")
-        )
-        self.data = {
-            "email": "test_email@example.com",
-            "password": "TestPassword123",
-        }
-        self.user = create_user(
-            first_name="first_name",
-            last_name="last_name",
-            email=self.data["email"],
-            password=self.data["password"],
-            is_active=True,
-        )
-        self.profile = create_student_profile(profile=create_profile(user=self.user))
-        self.user_2 = create_user(
-            first_name="first_name",
-            last_name="last_name",
-            email="test2@example.com",
-            password="Test12345",
-            is_active=True,
-        )
-        self.profile_2 = create_student_profile(
-            profile=create_profile(user=self.user_2)
-        )
-        self.lecturer_user_1 = create_user(
-            first_name="first_name",
-            last_name="last_name",
-            email="lecturer_1@example.com",
-            password=self.data["password"],
-            is_active=True,
-        )
-        self.lecturer_user_2 = create_user(
-            first_name="first_name",
-            last_name="last_name",
-            email="lecturer_2@example.com",
-            password=self.data["password"],
-            is_active=True,
-        )
-        self.lecturer_profile_1 = create_lecturer_profile(
-            profile=create_profile(user=self.lecturer_user_1, user_type="W")
-        )
-        self.lecturer_profile_2 = create_lecturer_profile(
-            profile=create_profile(user=self.lecturer_user_2, user_type="W")
-        )
-
-        self.technology_1 = create_technology(name="Python")
-        self.technology_2 = create_technology(name="JS")
-        self.technology_3 = create_technology(name="VBA")
-
-        self.lesson_1 = create_lesson(
-            title="Python lesson 1",
-            description="bbbb",
-            duration="90",
-            github_url="https://github.com/loopedupl/lesson",
-            price="9.99",
-            technologies=[self.technology_1],
-        )
-
-        self.lesson_2 = create_lesson(
-            title="Python lesson 2",
-            description="bbbb",
-            duration="30",
-            github_url="https://github.com/loopedupl/lesson",
-            price="2.99",
-            technologies=[self.technology_1],
-        )
-
-        self.module_1 = create_module(
-            title="Module 1", lessons=[self.lesson_1, self.lesson_2]
-        )
-
-        self.lesson_3 = create_lesson(
-            title="JS lesson 1",
-            description="bbbb",
-            duration="90",
-            github_url="https://github.com/loopedupl/lesson",
-            price="9.99",
-            technologies=[self.technology_2],
-        )
-
-        self.lesson_4 = create_lesson(
-            title="JS lesson 2",
-            description="bbbb",
-            duration="30",
-            github_url="https://github.com/loopedupl/lesson",
-            price="2.99",
-            technologies=[self.technology_2],
-        )
-
-        self.module_2 = create_module(
-            title="Module 2", lessons=[self.lesson_3, self.lesson_4]
-        )
-
-        self.lesson_5 = create_lesson(
-            title="VBA lesson 1",
-            description="bbbb",
-            duration="90",
-            github_url="https://github.com/loopedupl/lesson",
-            price="9.99",
-            technologies=[self.technology_3],
-        )
-
-        self.lesson_6 = create_lesson(
-            title="VBA lesson 2",
-            description="bbbb",
-            duration="30",
-            github_url="https://github.com/loopedupl/lesson",
-            price="2.99",
-            technologies=[self.technology_3],
-        )
-
-        self.module_3 = create_module(
-            title="Module 3", lessons=[self.lesson_5, self.lesson_6]
-        )
-
-    def test_title_filter(self):
-        # login
-        login(self, self.admin_data["email"], self.admin_data["password"])
-        self.assertTrue(auth.get_user(self.client).is_authenticated)
-        # get data
-        column = "title"
-        variable = str(self.module_1.title)
-        response = self.client.get(f"{self.endpoint}?{column}={variable}")
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        data = json.loads(response.content)
-        records_count = data["records_count"]
-        results = data["results"]
-        self.assertEqual(records_count, 1)
-        values = list(set([variable == record[column] for record in results]))
-        self.assertTrue(len(values) == 1)
-        self.assertTrue(values[0])
-
-
-class ModuleOrderTest(APITestCase):
-    def setUp(self):
-        self.endpoint = "/api/modules"
-        self.admin_data = {
-            "email": "admin_test_email@example.com",
-            "password": "TestPassword123",
-        }
-        self.admin_user = create_user(
-            first_name="first_name",
-            last_name="last_name",
-            email=self.admin_data["email"],
-            password=self.admin_data["password"],
-            is_active=True,
-            is_staff=True,
-        )
-        self.admin_profile = create_admin_profile(
-            profile=create_profile(user=self.admin_user, user_type="A")
-        )
-        self.data = {
-            "email": "test_email@example.com",
-            "password": "TestPassword123",
-        }
-        self.user = create_user(
-            first_name="first_name",
-            last_name="last_name",
-            email=self.data["email"],
-            password=self.data["password"],
-            is_active=True,
-        )
-        self.profile = create_student_profile(profile=create_profile(user=self.user))
-        self.user_2 = create_user(
-            first_name="first_name",
-            last_name="last_name",
-            email="test2@example.com",
-            password="Test12345",
-            is_active=True,
-        )
-        self.profile_2 = create_student_profile(
-            profile=create_profile(user=self.user_2)
-        )
-        self.lecturer_user_1 = create_user(
-            first_name="first_name",
-            last_name="last_name",
-            email="lecturer_1@example.com",
-            password=self.data["password"],
-            is_active=True,
-        )
-        self.lecturer_user_2 = create_user(
-            first_name="first_name",
-            last_name="last_name",
-            email="lecturer_2@example.com",
-            password=self.data["password"],
-            is_active=True,
-        )
-        self.lecturer_profile_1 = create_lecturer_profile(
-            profile=create_profile(user=self.lecturer_user_1, user_type="W")
-        )
-        self.lecturer_profile_2 = create_lecturer_profile(
-            profile=create_profile(user=self.lecturer_user_2, user_type="W")
-        )
-
-        self.technology_1 = create_technology(name="Python")
-        self.technology_2 = create_technology(name="JS")
-        self.technology_3 = create_technology(name="VBA")
-
-        self.lesson_1 = create_lesson(
-            title="Python lesson 1",
-            description="bbbb",
-            duration="90",
-            github_url="https://github.com/loopedupl/lesson",
-            price="9.99",
-            technologies=[self.technology_1],
-        )
-
-        self.lesson_2 = create_lesson(
-            title="Python lesson 2",
-            description="bbbb",
-            duration="30",
-            github_url="https://github.com/loopedupl/lesson",
-            price="2.99",
-            technologies=[self.technology_1],
-        )
-
-        self.module_1 = create_module(title="Module 1", lessons=[self.lesson_1])
-
-        self.lesson_3 = create_lesson(
-            title="JS lesson 1",
-            description="bbbb",
-            duration="90",
-            github_url="https://github.com/loopedupl/lesson",
-            price="9.99",
-            technologies=[self.technology_2],
-        )
-
-        self.lesson_4 = create_lesson(
-            title="JS lesson 2",
-            description="bbbb",
-            duration="30",
-            github_url="https://github.com/loopedupl/lesson",
-            price="2.99",
-            technologies=[self.technology_2],
-        )
-
-        self.module_2 = create_module(
-            title="Module 2", lessons=[self.lesson_3, self.lesson_4]
-        )
-
-        self.lesson_5 = create_lesson(
-            title="VBA lesson 1",
-            description="bbbb",
-            duration="90",
-            github_url="https://github.com/loopedupl/lesson",
-            price="9.99",
-            technologies=[self.technology_3],
-        )
-
-        self.lesson_6 = create_lesson(
-            title="VBA lesson 2",
-            description="bbbb",
-            duration="30",
-            github_url="https://github.com/loopedupl/lesson",
-            price="2.99",
-            technologies=[self.technology_3],
-        )
-
-        self.module_3 = create_module(
-            title="Module 3", lessons=[self.lesson_5, self.lesson_6, self.lesson_2]
-        )
-
-        self.fields = ["title", "lessons_count"]
-
-    def is_float(self, element: any) -> bool:
-        if element is None:
-            return False
-        try:
-            float(element)
-            return True
-        except ValueError:
-            return False
-
-    def test_ordering(self):
-        for field in self.fields:
-            login(self, self.admin_data["email"], self.admin_data["password"])
-            self.assertTrue(auth.get_user(self.client).is_authenticated)
-            # get data
-            response = self.client.get(f"{self.endpoint}?sort_by={field}")
-            self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-            data = json.loads(response.content)
-            count = data["records_count"]
-            results = data["results"]
-            self.assertEqual(count, 3)
-            if "_id" in field:
-                field1, field2 = field.split("_")
-                parent_objects = [
-                    course[field1] for course in results if course[field1] is not None
-                ]
-                field_values = [
-                    parent_object[field2] for parent_object in parent_objects
-                ]
-            elif "coupon_code" in field:
-                field1, field2 = field.split("_")
-                field_values = [course[field1][field2] for course in results]
-            elif "user_email" in field:
-                field1, field2 = field.split("_")
-                field_values = [course[field1][field2] for course in results]
-            elif "lessons_count" in field:
-                field_values = [len(course["lessons"]) for course in results]
-            else:
-                field_values = [course[field] for course in results]
-            if isinstance(field_values[0], dict):
-                self.assertEqual(
-                    field_values, sorted(field_values, key=lambda d: d["name"])
-                )
-            else:
-                field_values = [
-                    field_value
-                    if not self.is_float(field_value)
-                    else float(field_value)
-                    for field_value in field_values
-                ]
-                self.assertEqual(field_values, sorted(field_values))
-            # get data
-            response = self.client.get(f"{self.endpoint}?sort_by=-{field}")
-            self.assertEqual(response.status_code, status.HTTP_200_OK)
-            data = json.loads(response.content)
-            count = data["records_count"]
-            results = data["results"]
-            self.assertEqual(count, 3)
-            if "_id" in field:
-                field1, field2 = field.split("_")
-                parent_objects = [
-                    course[field1] for course in results if course[field1] is not None
-                ]
-                field_values = [
-                    parent_object[field2] for parent_object in parent_objects
-                ]
-            elif "coupon_code" in field:
-                field1, field2 = field.split("_")
-                field_values = [course[field1][field2] for course in results]
-            elif "user_email" in field:
-                field1, field2 = field.split("_")
-                field_values = [course[field1][field2] for course in results]
-            elif "lessons_count" in field:
-                field_values = [len(course["lessons"]) for course in results]
-            else:
-                field_values = [course[field] for course in results]
-            if isinstance(field_values[0], dict):
-                self.assertEqual(
-                    field_values,
-                    sorted(field_values, key=lambda d: d["name"], reverse=True),
-                )
-            else:
-                field_values = [
-                    field_value
-                    if not self.is_float(field_value)
-                    else float(field_value)
-                    for field_value in field_values
-                ]
-                self.assertEqual(field_values, sorted(field_values, reverse=True))
