+++ backend/wishlist/views.py
@@ -7,7 +7,7 @@ from profile.models import Profile
 
 class WishlistViewSet(ModelViewSet):
     http_method_names = ["get", "post", "delete"]
-    queryset = Wishlist.objects.all().select_related("student").order_by("id")
+    queryset = Wishlist.objects.all().select_related("student")
     serializer_class = WishlistGetSerializer
     permission_classes = [IsAuthenticated]
